chore: Create `author` command
Add a `author` command, the command lists the
contributors from a git repository.
For now the names are being printed to the stdout.

Errors are defined at src/errors/def.go
A user calls the error function with a number like 3
then the command exists with the error and set message.
The errors are in development stage.

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>