chore: Add `task` cmd for file and dir creation
The task command is resposible for creating the files
and directories as specified.
It uses the last part of the working directory to
figure out what directory/file to create.
For now strings are allowed as input.

No error checking is done as of now.
Logging is yet to be implemented.

If a user wants to create a directory, then they can use:
	--dir <id>
If a user wants to create a file, then they can use:
	--file <id>

`<id>` is defined as the first part before '-' in the ALX
structure.
For example in the repo `alx-low_level_programming` the directory
`0x08-recursion` has an id of `0x08` and the file
`0-puts_recursion.c` has an id of `0`.

The above structure is ideal for less typing more results.
Full directory names and Files are not allowed as of now
(no reason at all).

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>