fix: Use git remote instead of the cwd
Now the git remote is used instead of the current working directory.
This means that there must be an upstream for it to work.

Another assumption is that the remote is named 'origin'.
Although true for most repos out there, there are some that will
fail on this rule (a safe bet overall).

The `cwd` may be supported in the future but due to how repositories
are created/cloned it may be difficult.

An example:

```git
git clone https://github.com/gekkowrld/acli.git a
```

This will clone the repo to `a`.
Checking for the `cwd` in this case will fail and so the justification of using `remote`.
The `cwd` may be used as backup if there is no `remote` in the future iterations though.

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>